end of lesson 2

+++ 001 - OOPS Review/001 - OOPS Review/FencePanel.cs
@@ -89,7 +89,7 @@ namespace _001___OOPS_Review
 
 
         //Greddy
-        public FencePanel(int height, int width, string style, double? price)
+        public FencePanel(double height, double width, string style, double? price)
         {
             Height = height;
             Width = width;
